+++ _examples/sessions/securecookie/main.go
@@ -20,14 +20,13 @@ func newApp() *iris.Application {
 	cookieName := "_session_id"
 	// AES only supports key sizes of 16, 24 or 32 bytes.
 	// You either need to provide exactly that amount or you derive the key from what you type in.
-	hashKey := []byte("the-big-and-secret-fash-key-here")
-	blockKey := []byte("lot-secret-of-characters-big-too")
-	secureCookie := securecookie.New(hashKey, blockKey)
+	hashKey := securecookie.GenerateRandomKey(64)
+	blockKey := securecookie.GenerateRandomKey(32)
+	s := securecookie.New(hashKey, blockKey)
 
 	mySessions := sessions.New(sessions.Config{
 		Cookie:       cookieName,
-		Encode:       secureCookie.Encode,
-		Decode:       secureCookie.Decode,
+		Encoding:     s,
 		AllowReclaim: true,
 	})